backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
python-dotenv>=1.0
google-genai>=0.8
supabase>=2.0
easyocr>=1.7
Pillow>=10.0
numpy>=1.26
opencv-python-headless>=4.9
 